+++ Address-Finder/Address-Finder/Structs/APIResponse.swift
@@ -1,0 +1,12 @@
+//
+//  APIResponse.swift
+//  Address-Finder
+//
+//  Created by Daye on 2021/08/01.
+//
+
+import Foundation
+
+struct APIResponse: Codable {
+    let documents: [Documents]
+}
