+++ Address-Finder/Podfile
@@ -4,7 +4,6 @@
 
 target 'Address-Finder' do
   pod 'Alamofire', '~> 4.4'
-  pod 'SwiftyJSON', '~>4.0'
   # Comment the next line if you don't want to use dynamic frameworks
   use_frameworks!
